remove use of catchsegv (#586)
* remove use of catchsegv
* fixes for numpy 1.24 compatibility
* numpy<1.24 for compatibility with coremltools
* torch 1.13.1, torchvision 0.14.1
* skip onnx simplifier test

+++ src/openpifpaf/encoder/caf.py
@@ -190,7 +190,7 @@ class AssociationFiller:
         filled_ij = set()
         for f in frange:
             for xyo in xyv:
-                fij = np.round(joint1[:2] + f * offset + xyo).astype(np.int) + self.config.padding
+                fij = np.round(joint1[:2] + f * offset + xyo).astype(np.intc) + self.config.padding
                 if fij[0] < 0 or fij[0] >= self.field_shape[2] or \
                    fij[1] < 0 or fij[1] >= self.field_shape[1]:
                     continue
